update: refactor highlight behavior
Refactor the highlight behavior with proper structured files, names and classes

Ref: VRB-2933
