+++ Source/Core/Runtime/Behaviors/ColorHighlightBehaviorBase.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 6c7d1e74fb5eb994c900497a2db30816
