+++ Source/Core/Runtime/Behaviors/IColorHighlightBehaviorData.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: c299dda4eb1a1f44fb8094f94cd37f3e
