+++ Source/Core/Runtime/Behaviors/IHighlightingBehavior.cs
@@ -1,0 +1,9 @@
+namespace VRBuilder.Core.Behaviors
+{
+    /// <summary>
+    /// Marker interface for all behaviors that highlight scene objects.
+    /// </summary>
+    public interface IHighlightingBehavior : IBehavior
+    {
+    }
+}
