+++ Source/Core/Runtime/Behaviors/IHighlightingBehavior.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: eff61510f987dff4ea93d0c06614deb7
