+++ Source/Core/Runtime/Behaviors/IObjectHighlightBehavior.cs
@@ -1,0 +1,9 @@
+namespace VRBuilder.Core.Behaviors
+{
+    /// <summary>
+    /// Marker interface for object highlight behaviors.
+    /// </summary>
+    public interface IObjectHighlightBehavior : IHighlightingBehavior
+    {
+    }
+}
