+++ Source/Core/Runtime/Behaviors/IObjectHighlightBehavior.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: d43cf74c61ff7b244a781ed088ebf383
