+++ Source/Core/Runtime/Properties/ColorHighlightPropertyBase.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 401f5ae53ad40a64eb2101e610c1e5a8
